funciones factura y algunos cambios de models por las tablas y dependencias

+++ ProyectoDBP-TIENDA/Models/TemporalVenta.cs
@@ -4,7 +4,7 @@
     {
         public int IdPro { get; set; }
         public string DesPro { get; set; }
-        public double PrePro { get; set; }
+        public decimal PrePro { get; set; }
         public int StkAct { get; set; }
         public int cantidad { get; set; }
         public virtual ICollection<TbFactura> TbFacturas { get; set; } = new List<TbFactura>();
